+++ assets/tailwind.config.js
@@ -1,7 +1,3 @@
-module.exports = {
-  content: [
-    '/app/assets/js/**/*.js',
-    '/app/assets/css/**/*.css',
-    '/app/src/**/*.html'
-  ],
-}
+export default {
+  plugins: ["@tailwindcss/postcss"],
+};