+++ CHANGELOG.md
@@ -39,8 +39,9 @@ Changelog](https://keepachangelog.com/en/1.0.0/).
 #### Front-end dependencies
 
 - Update `esbuild` to `0.24.2`
-- Update `postcss` to `8.5.1`
-- Update `tailwindcss` to `3.4.17`
+- Update `tailwindcss` to `4.0.0`
+- Update `@tailwindcss/cli` to `4.0.0`
+- Update `@tailwindcss/postcss` to `4.0.0`
 
 ## [0.11.0] - 2024-08-09